% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir.R
\name{createNode}
\alias{createNode}
\title{Function to create some node}
\usage{
createNode(
  nodeType = c("Ridge"),
  otputDim = NULL,
  inputDim = NULL,
  name = NULL,
  ridge = 0,
  inputBias = TRUE,
  dtype = "float64"
)
}
\arguments{
\item{nodeType}{}

\item{otputDim}{Output dimension of the Node. Dimension of its state.}

\item{inputDim}{Input dimension of the Node.}

\item{name}{Name of the Node. It must be a unique identifier.}

\item{ridge}{float, default to 0.0 L2 regularization parameter.}

\item{inputBias}{bool, default to True If True, then a bias parameter will 
be learned along with output weights.}

\item{dtype}{Numerical type for node parameters}
}
\description{
Function to create some node
}
