% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{dfCovid}
\alias{dfCovid}
\title{Datagouv covid-19 dataset}
\format{
A data frame with 962 rows and 4 variables
}
\usage{
data(dfCovid)
}
\description{
A dataset containing the data from datagouv.fr concerning covid-19 infections in Aquitaine. Data related to hospitalizations can be found at Santé publique France  - Data downloaded at https://www.data.gouv.fr/fr/datasets/r/08c18e08-6780-452d-9b8c-ae244ad529b3, update from 26/01/2023. Data related to RT-PCR can be found at Santé publique France  - Data downloaded at https://www.data.gouv.fr/fr/datasets/r/10639654-3864-48ac-b024-d772c218c4c1, update from 26/01/2023.
}
\details{
\itemize{
  \item date. The date
  \item hosp. Number of person hospitalized with SARS-CoV-2 in Aquitaine.
  \item Positive. Number of person with a positive RT-PCR in Aquitaine.
  \item Tested. Number of person with a RT-PCR in Aquitaine.
}
}
\keyword{datasets}
