% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_numpy}
\alias{install_numpy}
\title{Install numpy}
\usage{
install_numpy(envname = "R-reticulate", method = "auto")
}
\arguments{
\item{envname}{\code{str} name of environment. Default is R-reticulate}

\item{method}{\code{str} type of environment type \code{(virtualenv, conda)}. 
Default is auto \code{(virtualenv is not available on Windows)}}
}
\description{
Install numpy
}
