% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_reservoirpy}
\alias{install_reservoirpy}
\title{Install reservoirpy}
\usage{
install_reservoirpy(envname = "r-reticulate", method = "auto")
}
\arguments{
\item{envname}{\code{str} name of environment. Default is R-reticulate}

\item{method}{\code{str} type of environment type \code{(virtualenv, conda)}. 
Default is auto \code{(virtualenv is not available on Windows)}}
}
\value{
A NULL object after installing reservoirpy python module.
}
\description{
Install reservoirpy
}
