% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_Reservoir.R
\name{node_Reservoir}
\alias{node_Reservoir}
\title{Create a reservoir network}
\usage{
node_Reservoir(units = NULL, lr = 1, sr = NULL, input_bias = TRUE, ...)
}
\arguments{
\item{units}{Number of reservoir units. If NULL, the number of units will be inferred from the W matrix shape.}

\item{lr}{Leaking rate (1.0 by default).}

\item{sr}{Spectral radius of W (optional).}

\item{input_bias}{If False, no bias is added to inputs.}

\item{...}{Additional parameters given to Reservoir from reservoirpy}
}
\description{
Create a reservoir network
}
\examples{
\dontrun{
node_Reservoir(units = as.integer(100), lr = 1.0, sr = 0.3)
}
}
