% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_Ridge.R
\name{node_Ridge}
\alias{node_Ridge}
\title{Ridge readout}
\usage{
node_Ridge(ridge = 0, input_bias = TRUE, ...)
}
\arguments{
\item{ridge}{float, default to \code{0.0}. L2 regularization parameter.}

\item{input_bias}{If True, then a bias parameter will be learned along with output weights.}

\item{...}{Additional arguments given to Ridge methods of reservoirpy}
}
\description{
Ridge readout
}
\examples{
reservoirpy <- reticulate::import("reservoirpy")
node_Ridge(ridge = 0.5)
}
