% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir.R
\name{nodes_operator}
\alias{nodes_operator}
\title{Operator \code{>>}}
\usage{
nodes_operator(node1, node2)
}
\arguments{
\item{node1}{(Node)}

\item{node2}{(Node)}
}
\value{
Node or list of nodes.
}
\description{
Take two nodes and applies \code{>>}
}
\examples{
\dontrun{
source = reservoir::createNode("Input")
reservoir = reservoir::createNode("Reservoir", units = 500, lr=0.1, sr=0.9)
nodes_operator(source, reservoir)
}

}
