% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{plot.reservoir_run}
\alias{plot.reservoir_run}
\title{plot.reservoir_run}
\usage{
\method{plot}{reservoir_run}(object, vec_nodes = c(1:20), vec_time = NULL)
}
\arguments{
\item{object}{A reservoir_run object}

\item{vec_nodes}{Number of nodes to plot}

\item{vec_time}{Time to plot}
}
\value{
A ggplot
}
\description{
plot.reservoir_run
}
