% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir.R
\name{predict_seq}
\alias{predict_seq}
\title{Run the Node forward function on a sequence of data.
Can update the state of the
Node several times}
\usage{
predict_seq(node, X, formState = NULL, stateFul = TRUE, reset = FALSE)
}
\arguments{
\item{node}{}

\item{X}{array-like of shape \code{([n_inputs], timesteps, input_dim)}
A sequence of data of shape (timesteps, features).}

\item{formState}{array of shape \code{(1, output_dim)}, optional 
Node state value to use at begining of computation.}

\item{stateFul}{\code{bool}, default to \code{TRUE} 
If True, Node state will be updated by this operation.}

\item{reset}{\code{bool}, default to \code{FALSE}
If True, Node state will be reset to zero before this operation.}
}
\description{
Run the Node forward function on a sequence of data.
Can update the state of the
Node several times
}
