% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rloguniform.R
\name{rloguniform}
\alias{rloguniform}
\title{rloguniform}
\usage{
rloguniform(n, min = 10^-1, max = 10^2)
}
\arguments{
\item{n}{number of sample}

\item{min}{minimum of the distribution}

\item{max}{maximum of the distribution}
}
\value{
A vector of simulated values
}
\description{
Simulate a log-uniform distribution
}
\examples{
rloguniform(n = 1)
}
