% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run the Node forward function on a sequence of data. Can update the state
of the Node several times.}
\usage{
run(node, X, from_state = NULL, stateful = TRUE, reset = FALSE)
}
\arguments{
\item{node}{node}

\item{X}{\code{array-like} of shape \code{[n_inputs], [series], timesteps, input_dim)}, optional
Input sequences dataset. If None, the method will try to fit
the parameters of the Node using the precomputed values returned
by previous call of :py:meth:\code{partial_fit}.}

\item{from_state}{\code{array-like} of shape \code{(1, output_dim)}, optional
Node state value to use at begining of computation}

\item{stateful}{\code{bool}, default to TRUE
If True, Node state will be updated by this operation.}

\item{reset}{\code{bool}, default to FALSE
If True, Node state will be reset to zero before this operation.}
}
\description{
Run the Node forward function on a sequence of data. Can update the state
of the Node several times.
}
