% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{summary.reservoir_run}
\alias{summary.reservoir_run}
\title{summary.reservoir_run}
\usage{
\method{summary}{reservoir_run}(object)
}
\arguments{
\item{object}{A reservoir_run object}
}
\value{
A dataframe with node activation
}
\description{
summary.reservoir_run
}
