% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fit.R
\name{summary.reservoirR_fit}
\alias{summary.reservoirR_fit}
\title{reservoirR_fit summary}
\usage{
\method{summary}{reservoirR_fit}(object, ...)
}
\arguments{
\item{object}{an object of class \code{reservoirR_fit} to summarized.}

\item{...}{further arguments.}
}
\value{
a \code{list} object
}
\description{
summary S3 method for reservoirR_fit object
}
\examples{
if(reticulate::py_module_available("reservoirpy")){
}

}
