% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir.R
\name{createNode}
\alias{createNode}
\title{Function to create some node}
\usage{
createNode(
  nodeType = c("Ridge"),
  units = NULL,
  lr = 1,
  sr = NULL,
  otputDim = NULL,
  inputDim = NULL,
  name = NULL,
  ridge = 0,
  inputBias = TRUE,
  dtype = "float64",
  ...
)
}
\arguments{
\item{nodeType}{Type of node. Default is \code{"Ridge"}.}

\item{units}{(int) optional 
Number of reservoir units. If None, the number of units will be infered from
the \code{W} matrix shape.}

\item{lr}{(float) default to 1.0
Neurons leak rate. Must be in :math:\code{[0, 1]}.}

\item{sr}{(float) optional
Spectral radius of recurrent weight matrix.}

\item{otputDim}{Output dimension of the Node. Dimension of its state.}

\item{inputDim}{Input dimension of the Node.}

\item{name}{Name of the Node. It must be a unique identifier.}

\item{ridge}{float, default to \code{0.0}. L2 regularization parameter.}

\item{inputBias}{bool, default to \code{TRUE}. If \code{TRUE}, then a bias parameter 
will be learned along with output weights.}

\item{dtype}{Numerical type for node parameters}

\item{...}{Others params}
}
\description{
Function to create some node
}
\examples{
if(interactive()){
readout <- createNode("Ridge")
}
}
