% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{dfCovid}
\alias{dfCovid}
\title{Bordeaux University Hospital covid-19 dataset}
\format{
A data frame with 612 rows and 4 variables
}
\usage{
data(dfCovid)
}
\description{
A dataset containing the Bordeaux Hospital covid-19 data (obfuscated < 10).
}
\details{
\itemize{
  \item date. The date
  \item hospitalisations. Number of person hospitalised in Bordeaux Hospital at a given date (obfuscated < 10)
  \item iptcc. The IPTCC index
  \item positive_pcr. Number of covid-19 positive RT-PCR in Bordeaux Hospital at a given date (obfuscated < 10)
}
}
\keyword{datasets}
