% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perf_22.R
\name{plot_perf_22}
\alias{plot_perf_22}
\title{plot_perf_22}
\usage{
plot_perf_22(x, y, dfPerf, perf_lab, trans = "log10")
}
\arguments{
\item{x}{The x feature}

\item{y}{The y feature}

\item{dfPerf}{The performance dataframe which should have the columns : perf, ridge, input_scaling, leaking_rate, spectral_radius. Where perf is the performance metric}

\item{perf_lab}{The label of the performance metric.}

\item{trans}{The transformation (default is "log10")}
}
\value{
A 2x2 plot
}
\description{
Unit plot for 2x2 function
}
