% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fit.R
\name{print.summary.reservoirR_fit}
\alias{print.summary.reservoirR_fit}
\title{reservoirR_fit print summary}
\usage{
\method{print}{summary.reservoirR_fit}(res, ...)
}
\arguments{
\item{res}{an object of class \code{summary.reservoirR_fit} to print.}

\item{...}{further arguments.}
}
\description{
print S3 method for summary.reservoirR_fit object
}
