% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_reservoir_predict_seq.R
\name{summary.reservoir_predict_seq}
\alias{summary.reservoir_predict_seq}
\title{summary.reservoir_predict_seq}
\usage{
\method{summary}{reservoir_predict_seq}(object)
}
\arguments{
\item{object}{A reservoir_predict_seq object}
}
\value{
A dataframe with node activation
}
\description{
summary.reservoir_predict_seq
}
